Add missing TP Link warning to Danger rules
In order to give other developers context when visiting our PR, we
want to link our pull request to specific Targetprocess tickets, where
the context/requirements of the changes are better described and may
have further context in the comments between devs and product.

This adds a Danger alert into the Pull Request when it does not include
a link to our Targetprocess domain.

The implementation is quite relaxed so either copy the URL from the
browser or copying the proper ticket link will both be accepted.
As long as our TP subdomain shows up in the PR description we can
assume that is a proper link to a ticket.